reformatted register page
